requirements.txt
streamlit==1.31.0
pandas==2.2.0
openai==1.12.0
google-generativeai>=0.8.0
python-dotenv==1.0.1
streamlit-mic-recorder==0.0.8
pydub==0.25.1
sqlparse==0.4.4

